One To Many Lecture

+++ EntityFrameworkLectures/Controllers/PostsController.cs
@@ -1,5 +1,6 @@
 using EntityFrameworkLectures.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 public class PostsController : Controller
 {
@@ -38,7 +39,7 @@ public class PostsController : Controller
     [HttpPost("/posts/create")]
     public IActionResult Create(Post newPost)
     {
-        if (!loggedIn)
+        if (!loggedIn || uid == null)
         {
             return RedirectToAction("Index", "Users");
         }
@@ -49,6 +50,10 @@ public class PostsController : Controller
             // allows that method to run w/o creating a new request response cycle (so we keep our errors)
             return New();
         }
+
+        // attaching currently logged in user's ID to the newPost
+        newPost.UserId = (int)uid;
+
         Console.WriteLine(newPost.PostId);
         // ModelState IS valid
         db.Posts.Add(newPost);
@@ -75,7 +80,11 @@ public class PostsController : Controller
         {
             return RedirectToAction("Index", "Users");
         }
-        List<Post> allPosts = db.Posts.ToList();
+
+        // get all posts, and include author info for each individual post in addition to the foreign key
+        List<Post> allPosts = db.Posts.Include(p => p.Author).ToList();
+            // ^ what you would have to write in SQL
+        // SELECT * FROM posts JOIN users on post.UserId = user.UserId
 
         return View("All", allPosts);
     }
@@ -87,7 +96,7 @@ public class PostsController : Controller
         {
             return RedirectToAction("Index", "Users");
         }
-        Post? post = db.Posts.FirstOrDefault(p => p.PostId == onePostId);
+        Post? post = db.Posts.Include(p => p.Author).FirstOrDefault(p => p.PostId == onePostId);
 
         // In case user manually types in an invalid ID in the url
         if (post == null)
@@ -107,7 +116,7 @@ public class PostsController : Controller
         }
         Post? post = db.Posts.FirstOrDefault(p => p.PostId == deletedPostId);
 
-        if (post != null)
+        if (post != null && post.UserId == uid)
         {
             db.Posts.Remove(post);
             db.SaveChanges();
@@ -124,10 +133,11 @@ public class PostsController : Controller
         }
         Post? post = db.Posts.FirstOrDefault(p => p.PostId == postId);
 
-        if(post == null)
+        if(post == null || post.UserId != uid)
         {
             return RedirectToAction("All");
         }
+
         return View("Edit", post);
     }
 
@@ -145,7 +155,7 @@ public class PostsController : Controller
 
         Post? dbPost = db.Posts.FirstOrDefault(post => post.PostId == postId);
 
-        if (dbPost == null)
+        if (dbPost == null || dbPost.UserId != uid)
         {
             return RedirectToAction("All");
         }
